advanced_python_calculator/requirements.txt
pandas
pylint
pytest
pytest-cov
python-dotenv  # for environment variables
